+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Improvements
 Bugfixes
 --------
 - Fixed an issue when trying to read hdf5 metadata from non-hdf5 files.
+- Fixed in issue in Parameter where sub-type checking in tuple/set/list Parameters
+  was not enforced.
 
 
 v24.09.19